rename demo project directory reformat mainwindows.cs
